+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 0.3.0 (in progress)
 -------------------
 
+- Added Layer7 encryption
 - Added PayPal endpoint
 - Added Transfer endpoint
 - Added Payment Status Transitions endpoint